fix: gitalk cannot login; feat: manifest
